always store the auth token in local storage
Previously, auth token was stored in the session storage by default (local storage was used in case `remember me` was ticked). This default behaviour forced the user to always sign in while opening a new tab of the application.
By always storing the token in the local storage, this issue is solved.
Now, `remember me` affects only the lifespan of the token.